fix(storybook): improve and simplify e2e tests
