--- e2e/storybook-angular/jest.config.ts
@@ -1,10 +1,0 @@
-export default {
-  transform: {
-    '^.+\\.[tj]sx?$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }],
-  },
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'html'],
-  maxWorkers: 1,
-  globals: {},
-  displayName: 'e2e-storybook-angular',
-  preset: '../../jest.preset.js',
-};
